requirements.txt
numpy
scikit-learn
torch
torchvision
pytorch3d
trimesh
imageio
matplotlib
Pillow
